pyproject.toml
[project]
name = "ro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.11.0",
    "django-jazzmin>=3.0.1",
    "django-silk>=5.4.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "ruff>=0.14.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]

# ===============================
# BLACK (FORMATTER ONLY)
# ===============================
[tool.black]
line-length = 88
target-version = ["py313"]

# ===============================
# RUFF - LINT + AUTOFIX
# ===============================
[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["migrations", ".venv", "env"]
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "W", "D", "UP", "B", "DJ"]
ignore = [
    "D100", 
    "D101", 
    "D102", 
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D212",
]

# ===============================
# RUFF FORMAT CONFIG
# ===============================
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# ===============================
# RUFF ISORT (IMPORT MANAGEMENT)
# ===============================
[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
known-first-party = ["backend", "api", "users"]


